tests/maintenance: separate utility mixin
separate maintenance utilities into a mixin so that we can trigger
maintenance mode in other tests where needed
